feat: add blink-core
