--- blink-api/src/main/kotlin/com/jordyma/blink/common/entity/BaseEntity.kt
@@ -1,30 +1,0 @@
-package com.jordyma.blink.common.entity
-
-import jakarta.persistence.Column
-import jakarta.persistence.EntityListeners
-import jakarta.persistence.MappedSuperclass
-import lombok.Getter
-import org.springframework.data.annotation.CreatedDate
-import org.springframework.data.annotation.LastModifiedDate
-import org.springframework.data.jpa.domain.support.AuditingEntityListener
-import java.time.LocalDateTime
-
-@Getter
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener::class)
-open class BaseTimeEntity {
-
-    @CreatedDate
-    @Column(name = "created_at", updatable = false)
-    var createdAt: LocalDateTime? = null
-        private set
-
-    @LastModifiedDate
-    @Column(name = "updated_at")
-    var updatedAt: LocalDateTime? = null
-        private set
-
-    @Column(name = "deleted_at")
-    var deletedAt: LocalDateTime? = null
-        private set
-}
