--- blink-api/src/main/kotlin/com/jordyma/blink/feed/repository/FeedRepository.kt
@@ -1,9 +1,0 @@
-package com.jordyma.blink.feed.repository
-
-import com.jordyma.blink.feed.entity.Feed
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface FeedRepository : JpaRepository<Feed, Long>, FeedRepositoryCustom {
-
-}
-
