--- blink-api/src/main/kotlin/com/jordyma/blink/folder/repository/CustomFolderRepository.kt
@@ -1,11 +1,0 @@
-package com.jordyma.blink.folder.repository
-
-import com.jordyma.blink.folder.entity.Folder
-import com.jordyma.blink.user.entity.User
-
-interface CustomFolderRepository {
-
-    fun findAllByUser(user: User): List<Folder>
-
-    fun deleteFolder(folder: Folder): Unit
-}
