--- blink-api/src/main/kotlin/com/jordyma/blink/notice/entity/Notice.kt
@@ -1,19 +1,0 @@
-package com.jordyma.blink.notice.entity
-
-import com.jordyma.blink.common.entity.BaseTimeEntity
-import jakarta.persistence.*
-
-@Entity
-class Notice(
-
-    @Column(nullable = false, length = 50)
-    val title: String,
-
-    @Column(nullable = false, length = 1000)
-    val content: String,
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    val id: Long,
-
-): BaseTimeEntity()
