--- blink-api/src/main/kotlin/com/jordyma/blink/notice/repository/NoticeRepository.kt
@@ -1,8 +1,0 @@
-package com.jordyma.blink.notice.repository
-
-
-import com.jordyma.blink.notice.entity.Notice
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface NoticeRepository: JpaRepository<Notice, Long> {
-}
