--- blink-api/src/main/kotlin/com/jordyma/blink/user/entity/Role.kt
@@ -1,11 +1,0 @@
-package com.jordyma.blink.user.entity
-
-import org.springframework.security.core.GrantedAuthority
-
-enum class Role: GrantedAuthority {
-    ADMIN, USER;
-
-    override fun getAuthority(): String {
-        return name;
-    }
-}
