--- blink-api/src/main/kotlin/com/jordyma/blink/user/entity/SocialType.kt
@@ -1,6 +1,0 @@
-package com.jordyma.blink.user.entity
-
-enum class SocialType {
-    KAKAO,
-    APPLE,
-}
