+++ blink-core/src/main/kotlin/com.jordyma.blink/feed/vo/FeedDetailVo.kt
