+++ blink-core/src/main/kotlin/com.jordyma.blink/folder/Topic.kt
