+++ blink-core/src/main/kotlin/com.jordyma.blink/keyword/KeywordRepository.kt
@@ -1,0 +1,7 @@
+package com.jordyma.blink.keyword
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface KeywordRepository : JpaRepository<Keyword, Long> {
+    fun findByFeedId(feedId: Long): List<Keyword>
+}
