+++ blink-core/src/main/kotlin/com.jordyma.blink/recommend/Recommend.kt
@@ -1,0 +1,24 @@
+package com.jordyma.blink.folder.entity
+
+import com.jordyma.blink.common.BaseTimeEntity
+import com.jordyma.blink.feed.Feed
+import jakarta.persistence.*
+
+@Entity
+class Recommend (
+
+    @ManyToOne @JoinColumn(name = "feed_id")
+    var feed: Feed,
+
+    @Column(name = "folder_name", nullable = false)
+    val folderName: String,
+
+    @Column(name = "priority", nullable = false)
+    val priority: Int,
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Column(name = "id")
+    val id: Long? = null,
+
+    ): BaseTimeEntity()
