+++ blink-core/src/main/kotlin/com.jordyma.blink/user/Role.kt
@@ -1,0 +1,11 @@
+package com.jordyma.blink.user
+
+import org.springframework.security.core.GrantedAuthority
+
+enum class Role: GrantedAuthority {
+    ADMIN, USER;
+
+    override fun getAuthority(): String {
+        return name;
+    }
+}
