+++ blink-core/src/main/kotlin/com.jordyma.blink/user/SocialType.kt
@@ -1,0 +1,6 @@
+package com.jordyma.blink.user
+
+enum class SocialType {
+    KAKAO,
+    APPLE,
+}
