+++ settings.gradle.kts
@@ -2,3 +2,4 @@ rootProject.name = "blink"
 include("blink-api")
 include("blink-worker")
 include("blink-batch")
+include("blink-core")
